Speed up notebook tests by adding extra replacements file (#4077)
Provides the ability to speed up notebook execution tests by substituting smaller values for parameters that impact the runtime of the notebook.  To do this one supplies a file for a given `.ipynb` test with the same prefix but the `.tst` suffix.  This file contains replacements of the form `pattern->replacement`.  Some notes

* Suffix `.tst` chosen instead of `.test` because the later has the potential to make some grepping for tests less reliable and more noise prone
* All the patterns must be applied during the test. This avoids the common case of a pattern being a typo and not being applied.
* Open to different syntax than `->`.  Used it because it seemed unlikely to conflict with actual patterns and was indicative of what the replacement is actually doing.  

Notebook tests are taking up to 30 minutes in Google Actions.  This PR gets that down to around 11.  Nearly as fast as the Windows test :) 

Addresses #3603