[XEB] Multiprocessing spawn (#4090)
There are two multiprocessing "modes": fork and spawn. Fork used to always be the default. Now sometimes on Mac it defaults to spawn (?) but in any event: spawn works way better. This could prevent some of the c++-looking errors that @mrwojtek has seen. It fixed different-but-still-c++-looking-errors for me