Proto serialization v25 (#4333)
This PR creates a new class called CircuitSerializer that can serialize Circuits into protocol buffers more efficiently than existing mechanisms.

This class is similar to SerializableGateSet and op_(de)serializer, in that it serializes and deserializes cirq.Circuits into Program protocol buffers, but it is about 60% more space efficient and has significant time savings as well.

This serializer works by changing the Gate field and Args dict to a specific message for each allowed gate type.  It changes Qubits from a repeated Qubit message, to an index in the constant message.

This PR also adds a FloatArg message which is specific to Arguments where it is known the value will be a float (or a sympy expression).