Simultaneous readout (#4307)
Adds an experiment for simultaneous (parallel) readout.

This experiment (estimate_parallel_single_qubit_readout_errors)  initializes a number of bitstrings(trials) and initializes the qubits to those bitstrings and immediately measures them It then estimates zero state and one state errors averaged over all of the trials where the given qubit was initialized to |0> or |1> respectively.  

This change also collapses the isolated single qubit readout experiment to use this function as well.