Fix resolving integer divisions (#4521)
The cirq parameter resolver internally used `np.power` to evaluate `a/b`. This fails for integer values of `a` and `b`. This PR casts to a float, so that the following is allowed:
```
import cirq
import sympy

Rx_gate = cirq.rx(rads=sympy.Symbol('a')/sympy.Symbol('b'))
cirq.resolve_parameters(Rx_gate,{'a':1, 'b':2})
```
Fixes #4516