Add `cirq_google.FSimGateFamily` class. (#4586)
This PR adds a new `cirq_google.FSimGateFamily` class which can be used to:

1. Convert any compatible FSimGate type to another compatible FSimGate type.
2. Accept (potentially parameterized) instances of compatible FSimGate types.

The gate family is also used to replace existing logic in cirq_google for checking / converting against fsim types. Once merged, this would also be useful to perform similar checks and conversions in pyle. 

This is part of the roadmap item: #3243