Add link to dev environment page in CONTRIBUTING.md (#4843)
Might be helpful to add a section in CONTRIBUTING.md to describe dev environment setup as it could be one of the first docs contributors look at. WDYT?

edit: Just found links to development.md elsewhere in the doc. Do you think it makes sense to have a dedicated section, to make it obvious where to look for dev environment setup?