Replace AbstractEngineProcessorShim (#4842)
We were using `AbstractEngineProcessorShim` pending #4644 and friends. Those PRs have landed. For data hygiene as part of the workflow tools, we need to record what processor was used. This PR introduces an abstract `ProcessorRecord` interface and three implementations for mapping (processor_id, whether it's a simulator, what sort of noise) tuples to an AbstractProcessor instance for doing the work.

 - Remove `AbstractEngineProcessorShim`
 - Add dataclasses for data hygene
 - Now we can have proper serialization tests for anything that saves a `QuantumRuntimeConfiguration`, hence the new json files
 - We don't need to shim over `cirq.read_json` in tests that make use of `AbstractEngineProcessorShim` any more
 - Side effect: the new engine-processor-backed instances mean we have to use "real" qubits in the tests, so there are some knock-on test changes.

cc @dstrain115 @MichaelBroughton