The number of exponents is too damn high (#5456)
Just going to throw this out there for comment.

`assist_consistent_protocols` is used in a huge number of tests (many of which are themselves parameterized tests).  We test all of the protocols for a large number of different `pow`ed versions of the gate.  This ends up being something like 10% or our test time.  

I've reduced this number down to what I think is a bit more reasonable, (0, itself, inverse, a non-zero float, a symbol).