Remove Rigetti from local checks (#5632)
This remove the rigetti integration tests from the default of `pytest-and-incremental-coverage` and hence of `all`.  It keeps these tests in the CI workflow.  

The alternative to this would be to add the docker up and down to these scripts, but I'm hesitant to do this as these download external docker images.

Fixes #4660