Roll back the default on unitary to np.complex64, change default for final_state_vector (#5636)
Rolls back change in #5426 that set the default precision on `Circuit#unitary` to `np.complex128`.

Also sets the default on `Circuit#final_state_vector` to `np.complex128`.  There was a deprecation warning that this was going to be downgraded to `np.complex64`,, but this will now remain `np.complex128`.  This is breaking in that is someone explicitly passed `dtype=None` to this method it would now break.  

Rolls back #5426 #5616 #5537 #5535